pyproject.toml
[project]
name = "mem0-task"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bs4>=0.0.2",
    "dotenv>=0.9.9",
    "ghapi>=1.0.6",
    "github-dependents-info>=1.6.3",
    "google-generativeai>=0.8.4",
    "mem0ai>=0.1.86",
    "requests>=2.32.3",
]
